SpreadsheetEngineHateosHandlerTestCase2.engineContext(engine, metadata) extracted
